ajustes finais

+++ Modules/OrdemServico/Database/Migrations/2019_10_12_141202_telefone_table.php
@@ -13,11 +13,10 @@ class TelefoneTable extends Migration
      */
     public function up()
     {
-        
         Schema::create('telefone', function(Blueprint $table)
 		{
             $table->increments('id');
-            $table->integer('numero');
+            $table->bigInteger('numero');
             $table->integer('solicitante_id')->unsigned();
             $table->foreign('solicitante_id')->references('id')->on('solicitante');
         });
